Update updateannuallanduse to be consistent with its standalone version
Precision changed to double for updateannuallanduse_v2.c, and also
corresponding calls and arrays in glm2iac_mod.F90 (this eliminates
the extra float arrays that had to be created for the float version).
Rounding outputs to integer percent replaced with rounding to 1E-5
percent for pfts and also for harvest. Additional checks added to
ensure the exact same outputs across machines, compilers, and compile
optimizations. This has been tested on cori, compy, and mac osx,
using intel and gcc compilers. Final 4-way test across three machines
included a different compiler on each of cori and compy, both with
level 2 optimization, and the gcc compiler on mac osx with level 0
and level 3 optimization. The outputs are all identical for the
historical and the one tested future scenario (ssp5 rcp5).

Some minor bug fixes, and some changes to the reference year change
calculations. The previous year is the reference for all years after
2000, and pasture is tracked so that it isn't unnecessarily removed
by crops and so it is not "added" on top of existing pasture.

The land conversion assumptions have been changed to operate on
proportional addition and removal of pfts because this seems to
generate results more consistent with the current ssp5 rcp85.
Looking at the other scenarios indicates that these settings should
also work for those as well (although no others have been tested yet).

This is now the same exact file as for elm tools, but the standalone
flag is not defined so that it works within iESM.

 Changes to be committed:
	modified:   iac/bld/giac.template
	modified:   iac/bld/iac_glm_namelist_items.txt
	modified:   iac/coupling/glm2iac_mod.F90
	modified:   iac/coupling/unused/updateannuallanduse_interface.c.double
	modified:   iac/coupling/updateannuallanduse_interface.c
	modified:   iac/coupling/updateannuallanduse_v2.c